pallet configs exported
