+++ evm-template/runtime/src/lib.rs
@@ -7,79 +7,40 @@
 include!(concat!(env!("OUT_DIR"), "/wasm_binary.rs"));
 
 pub mod apis;
+pub mod configs;
 pub mod constants;
 pub mod governance;
 mod weights;
 pub mod xcm_config;
 
-#[cfg(feature = "async-backing")]
-use cumulus_pallet_parachain_system::RelayNumberMonotonicallyIncreases;
-#[cfg(not(feature = "async-backing"))]
-use cumulus_pallet_parachain_system::RelayNumberStrictlyIncreases;
-use cumulus_primitives_core::{AggregateMessageOrigin, ParaId};
 use fp_account::EthereumSignature;
 use frame_support::{
-    construct_runtime, derive_impl,
-    dispatch::DispatchClass,
-    parameter_types,
-    traits::{
-        AsEnsureOriginWithArg, ConstU32, ConstU64, Contains, EitherOfDiverse, FindAuthor,
-        InstanceFilter, TransformOrigin,
-    },
-    weights::{
-        ConstantMultiplier, Weight, WeightToFeeCoefficient, WeightToFeeCoefficients,
-        WeightToFeePolynomial,
-    },
-    PalletId,
+    construct_runtime, parameter_types,
+    traits::{ConstU32, EitherOfDiverse},
+    weights::{WeightToFeeCoefficient, WeightToFeeCoefficients, WeightToFeePolynomial},
 };
-use frame_system::{
-    limits::{BlockLength, BlockWeights},
-    EnsureRoot, EnsureSigned,
-};
-use governance::{
-    origins::{pallet_custom_origins, Treasurer},
-    TreasurySpender,
-};
-use pallet_ethereum::PostLogContent;
-use pallet_evm::{EVMCurrencyAdapter, EnsureAccountId20, IdentityAddressMapping};
+use frame_system::{EnsureRoot, EnsureSigned};
+use governance::origins::pallet_custom_origins;
 use pallet_xcm::{EnsureXcm, IsVoiceOfBody};
-use parachains_common::message_queue::{NarrowOriginToSibling, ParaIdToSibling};
-use parity_scale_codec::{Decode, Encode, MaxEncodedLen};
-use polkadot_runtime_common::{
-    impls::{LocatableAssetConverter, VersionedLocatableAsset, VersionedLocationConverter},
-    xcm_sender::NoPriceForMessageDelivery,
-    BlockHashCount, SlowAdjustingFeeUpdate,
-};
-use scale_info::TypeInfo;
 use smallvec::smallvec;
 pub use sp_consensus_aura::sr25519::AuthorityId as AuraId;
-use sp_core::{H160, U256};
+use sp_core::H160;
 #[cfg(any(feature = "std", test))]
 pub use sp_runtime::BuildStorage;
 use sp_runtime::{
     generic, impl_opaque_keys,
     traits::{
-        AccountIdLookup, BlakeTwo256, DispatchInfoOf, Dispatchable, IdentifyAccount,
-        IdentityLookup, PostDispatchInfoOf, Verify,
+        BlakeTwo256, DispatchInfoOf, Dispatchable, IdentifyAccount, PostDispatchInfoOf, Verify,
     },
     transaction_validity::{TransactionValidity, TransactionValidityError},
-    ConsensusEngineId, RuntimeDebug,
 };
 pub use sp_runtime::{MultiAddress, Perbill, Permill};
-use sp_std::{marker::PhantomData, prelude::*};
+use sp_std::prelude::*;
 #[cfg(feature = "std")]
 use sp_version::NativeVersion;
-use sp_version::RuntimeVersion;
-// XCM Imports
-use xcm::{
-    latest::{prelude::BodyId, InteriorLocation, Junction::PalletInstance},
-    VersionedLocation,
-};
-use xcm_builder::PayOverXcm;
-#[cfg(not(feature = "runtime-benchmarks"))]
-use xcm_builder::ProcessXcmMessage;
 
 pub use crate::{
+    configs::StakingAdminBodyId,
     constants::{
         currency::{deposit, CENTS, EXISTENTIAL_DEPOSIT, MICROCENTS, MILLICENTS},
         AVERAGE_ON_INITIALIZE_RATIO, BLOCK_PROCESSING_VELOCITY, DAYS, HOURS, MAXIMUM_BLOCK_WEIGHT,
@@ -205,6 +166,13 @@ pub type Executive = frame_executive::Executive<
     AllPalletsWithSystem,
 >;
 
+type ConsensusHook = cumulus_pallet_aura_ext::FixedVelocityConsensusHook<
+    Runtime,
+    RELAY_CHAIN_SLOT_DURATION_MILLIS,
+    BLOCK_PROCESSING_VELOCITY,
+    UNINCLUDED_SEGMENT_CAPACITY,
+>;
+
 /// Handles converting a weight scalar to a fee value, based on the scale and
 /// granularity of the node's balance type.
 ///
@@ -272,655 +240,12 @@ pub fn native_version() -> NativeVersion {
     NativeVersion { runtime_version: VERSION, can_author_with: Default::default() }
 }
 
-parameter_types! {
-    pub const Version: RuntimeVersion = VERSION;
-
-    // This part is copied from Substrate's `bin/node/runtime/src/lib.rs`.
-    //  The `RuntimeBlockLength` and `RuntimeBlockWeights` exist here because the
-    // `DeletionWeightLimit` and `DeletionQueueDepth` depend on those to parameterize
-    // the lazy contract deletion.
-    pub RuntimeBlockLength: BlockLength =
-        BlockLength::max_with_normal_ratio(MAX_BLOCK_LENGTH, NORMAL_DISPATCH_RATIO);
-    pub RuntimeBlockWeights: BlockWeights = BlockWeights::builder()
-        .base_block(BlockExecutionWeight::get())
-        .for_class(DispatchClass::all(), |weights| {
-            weights.base_extrinsic = ExtrinsicBaseWeight::get();
-        })
-        .for_class(DispatchClass::Normal, |weights| {
-            weights.max_total = Some(NORMAL_DISPATCH_RATIO * MAXIMUM_BLOCK_WEIGHT);
-        })
-        .for_class(DispatchClass::Operational, |weights| {
-            weights.max_total = Some(MAXIMUM_BLOCK_WEIGHT);
-            // Operational transactions have some extra reserved space, so that they
-            // are included even if block reached `MAXIMUM_BLOCK_WEIGHT`.
-            weights.reserved = Some(
-                MAXIMUM_BLOCK_WEIGHT - NORMAL_DISPATCH_RATIO * MAXIMUM_BLOCK_WEIGHT
-            );
-        })
-        .avg_block_initialization(AVERAGE_ON_INITIALIZE_RATIO)
-        .build_or_panic();
-    pub const SS58Prefix: u16 = 42;
-}
-
-pub struct NormalFilter;
-impl Contains<RuntimeCall> for NormalFilter {
-    fn contains(c: &RuntimeCall) -> bool {
-        match c {
-            // We filter anonymous proxy as they make "reserve" inconsistent
-            // See: https://github.com/paritytech/substrate/blob/37cca710eed3dadd4ed5364c7686608f5175cce1/frame/proxy/src/lib.rs#L270
-            RuntimeCall::Proxy(method) => !matches!(
-                method,
-                pallet_proxy::Call::create_pure { .. } | pallet_proxy::Call::kill_pure { .. }
-            ),
-            _ => true,
-        }
-    }
-}
-
-/// The default types are being injected by [`derive_impl`](`frame_support::derive_impl`) from
-/// [`ParaChainDefaultConfig`](`struct@frame_system::config_preludes::ParaChainDefaultConfig`),
-/// but overridden as needed.
-#[derive_impl(frame_system::config_preludes::ParaChainDefaultConfig as frame_system::DefaultConfig)]
-impl frame_system::Config for Runtime {
-    /// The data to be stored in an account.
-    type AccountData = pallet_balances::AccountData<Balance>;
-    /// The identifier used to distinguish between accounts.
-    type AccountId = AccountId;
-    /// The basic call filter to use in dispatchable.
-    type BaseCallFilter = NormalFilter;
-    /// The block type.
-    type Block = Block;
-    /// Maximum number of block number to block hash mappings to keep (oldest pruned first).
-    type BlockHashCount = BlockHashCount;
-    /// The maximum length of a block (in bytes).
-    type BlockLength = RuntimeBlockLength;
-    /// Block & extrinsics weights: base values and limits.
-    type BlockWeights = RuntimeBlockWeights;
-    /// The weight of database operations that the runtime can invoke.
-    type DbWeight = RocksDbWeight;
-    /// The type for hashing blocks and tries.
-    type Hash = Hash;
-    /// The lookup mechanism to get account ID from whatever is passed in
-    /// dispatchers.
-    type Lookup = AccountIdLookup<AccountId, ()>;
-    /// The maximum number of consumers allowed on a single account.
-    type MaxConsumers = ConstU32<16>;
-    /// The index type for storing how many extrinsics an account has signed.
-    type Nonce = Nonce;
-    /// The action to take on a Runtime Upgrade
-    type OnSetCode = cumulus_pallet_parachain_system::ParachainSetCode<Self>;
-    /// Converts a module to an index of this module in the runtime.
-    type PalletInfo = PalletInfo;
-    /// The aggregated dispatch type that is available for extrinsics.
-    type RuntimeCall = RuntimeCall;
-    /// The ubiquitous event type.
-    type RuntimeEvent = RuntimeEvent;
-    /// The ubiquitous origin type.
-    type RuntimeOrigin = RuntimeOrigin;
-    /// This is used as an identifier of the chain. 42 is the generic substrate prefix.
-    type SS58Prefix = SS58Prefix;
-    /// Runtime version.
-    type Version = Version;
-}
-
-parameter_types! {
-    pub MaximumSchedulerWeight: frame_support::weights::Weight = Perbill::from_percent(80) *
-        RuntimeBlockWeights::get().max_block;
-    pub const MaxScheduledPerBlock: u32 = 50;
-    pub const NoPreimagePostponement: Option<u32> = Some(10);
-}
-
-impl pallet_scheduler::Config for Runtime {
-    type MaxScheduledPerBlock = MaxScheduledPerBlock;
-    type MaximumWeight = MaximumSchedulerWeight;
-    type OriginPrivilegeCmp = frame_support::traits::EqualPrivilegeOnly;
-    type PalletsOrigin = OriginCaller;
-    type Preimages = Preimage;
-    type RuntimeCall = RuntimeCall;
-    type RuntimeEvent = RuntimeEvent;
-    type RuntimeOrigin = RuntimeOrigin;
-    type ScheduleOrigin = EnsureRoot<AccountId>;
-    type WeightInfo = pallet_scheduler::weights::SubstrateWeight<Runtime>;
-}
-
-parameter_types! {
-    pub const PreimageBaseDeposit: Balance = deposit(2, 64);
-    pub const PreimageByteDeposit: Balance = deposit(0, 1);
-    pub const PreimageHoldReason: RuntimeHoldReason = RuntimeHoldReason::Preimage(pallet_preimage::HoldReason::Preimage);
-}
-
-impl pallet_preimage::Config for Runtime {
-    type Consideration = frame_support::traits::fungible::HoldConsideration<
-        AccountId,
-        Balances,
-        PreimageHoldReason,
-        frame_support::traits::LinearStoragePrice<
-            PreimageBaseDeposit,
-            PreimageByteDeposit,
-            Balance,
-        >,
-    >;
-    type Currency = Balances;
-    type ManagerOrigin = EnsureRoot<AccountId>;
-    type RuntimeEvent = RuntimeEvent;
-    type WeightInfo = pallet_preimage::weights::SubstrateWeight<Runtime>;
-}
-
-impl pallet_timestamp::Config for Runtime {
-    #[cfg(feature = "experimental")]
-    type MinimumPeriod = ConstU64<0>;
-    #[cfg(not(feature = "experimental"))]
-    type MinimumPeriod = ConstU64<{ SLOT_DURATION / 2 }>;
-    /// A timestamp: milliseconds since the unix epoch.
-    type Moment = u64;
-    type OnTimestampSet = Aura;
-    type WeightInfo = pallet_timestamp::weights::SubstrateWeight<Runtime>;
-}
-
-impl pallet_authorship::Config for Runtime {
-    type EventHandler = (CollatorSelection,);
-    type FindAuthor = pallet_session::FindAccountFromAuthorIndex<Self, Aura>;
-}
-
-parameter_types! {
-    pub const MaxProxies: u32 = 32;
-    pub const MaxPending: u32 = 32;
-    pub const ProxyDepositBase: Balance = deposit(1, 40);
-    pub const AnnouncementDepositBase: Balance = deposit(1, 48);
-    pub const ProxyDepositFactor: Balance = deposit(0, 33);
-    pub const AnnouncementDepositFactor: Balance = deposit(0, 66);
-}
-
-/// The type used to represent the kinds of proxying allowed.
-/// If you are adding new pallets, consider adding new ProxyType variant
-#[derive(
-    Copy,
-    Clone,
-    Decode,
-    Default,
-    Encode,
-    Eq,
-    MaxEncodedLen,
-    Ord,
-    PartialEq,
-    PartialOrd,
-    RuntimeDebug,
-    TypeInfo,
-)]
-pub enum ProxyType {
-    /// Allows to proxy all calls
-    #[default]
-    Any,
-    /// Allows all non-transfer calls
-    NonTransfer,
-    /// Allows to finish the proxy
-    CancelProxy,
-    /// Allows to operate with collators list (invulnerables, candidates, etc.)
-    Collator,
-}
-
-impl InstanceFilter<RuntimeCall> for ProxyType {
-    fn filter(&self, c: &RuntimeCall) -> bool {
-        match self {
-            ProxyType::Any => true,
-            ProxyType::NonTransfer => !matches!(c, RuntimeCall::Balances { .. }),
-            ProxyType::CancelProxy => matches!(
-                c,
-                RuntimeCall::Proxy(pallet_proxy::Call::reject_announcement { .. })
-                    | RuntimeCall::Multisig { .. }
-            ),
-            ProxyType::Collator => {
-                matches!(c, RuntimeCall::CollatorSelection { .. } | RuntimeCall::Multisig { .. })
-            }
-        }
-    }
-}
-
-impl pallet_proxy::Config for Runtime {
-    type AnnouncementDepositBase = AnnouncementDepositBase;
-    type AnnouncementDepositFactor = AnnouncementDepositFactor;
-    type CallHasher = BlakeTwo256;
-    type Currency = Balances;
-    type MaxPending = MaxPending;
-    type MaxProxies = MaxProxies;
-    type ProxyDepositBase = ProxyDepositBase;
-    type ProxyDepositFactor = ProxyDepositFactor;
-    type ProxyType = ProxyType;
-    type RuntimeCall = RuntimeCall;
-    type RuntimeEvent = RuntimeEvent;
-    type WeightInfo = pallet_proxy::weights::SubstrateWeight<Runtime>;
-}
-
-parameter_types! {
-    pub const ExistentialDeposit: Balance = EXISTENTIAL_DEPOSIT;
-    pub const MaxFreezes: u32 = 0;
-    pub const MaxLocks: u32 = 50;
-    pub const MaxReserves: u32 = 50;
-}
-
-impl pallet_balances::Config for Runtime {
-    type AccountStore = System;
-    /// The type for recording an account's balance.
-    type Balance = Balance;
-    type DustRemoval = ();
-    type ExistentialDeposit = ExistentialDeposit;
-    type FreezeIdentifier = ();
-    type MaxFreezes = MaxFreezes;
-    type MaxLocks = MaxLocks;
-    type MaxReserves = MaxReserves;
-    type ReserveIdentifier = [u8; 8];
-    /// The ubiquitous event type.
-    type RuntimeEvent = RuntimeEvent;
-    type RuntimeFreezeReason = RuntimeFreezeReason;
-    type RuntimeHoldReason = RuntimeHoldReason;
-    type WeightInfo = pallet_balances::weights::SubstrateWeight<Runtime>;
-}
-
-parameter_types! {
-    pub const AssetDeposit: Balance = 10 * CENTS;
-    pub const AssetAccountDeposit: Balance = deposit(1, 16);
-    pub const ApprovalDeposit: Balance = MILLICENTS;
-    pub const StringLimit: u32 = 50;
-    pub const MetadataDepositBase: Balance = deposit(1, 68);
-    pub const MetadataDepositPerByte: Balance = deposit(0, 1);
-    pub const RemoveItemsLimit: u32 = 1000;
-}
-
-impl pallet_assets::Config for Runtime {
-    type ApprovalDeposit = ApprovalDeposit;
-    type AssetAccountDeposit = AssetAccountDeposit;
-    type AssetDeposit = AssetDeposit;
-    type AssetId = u32;
-    type AssetIdParameter = parity_scale_codec::Compact<u32>;
-    type Balance = Balance;
-    #[cfg(feature = "runtime-benchmarks")]
-    type BenchmarkHelper = ();
-    type CallbackHandle = ();
-    type CreateOrigin = AsEnsureOriginWithArg<EnsureSigned<AccountId>>;
-    type Currency = Balances;
-    type Extra = ();
-    type ForceOrigin = EnsureRoot<AccountId>;
-    type Freezer = ();
-    type MetadataDepositBase = MetadataDepositBase;
-    type MetadataDepositPerByte = MetadataDepositPerByte;
-    type RemoveItemsLimit = RemoveItemsLimit;
-    type RuntimeEvent = RuntimeEvent;
-    type StringLimit = StringLimit;
-    type WeightInfo = pallet_assets::weights::SubstrateWeight<Runtime>;
-}
-
-parameter_types! {
-    /// Relay Chain `TransactionByteFee` / 10
-    pub const TransactionByteFee: Balance = 10 * MICROCENTS;
-    pub const OperationalFeeMultiplier: u8 = 5;
-}
-
-impl pallet_transaction_payment::Config for Runtime {
-    type FeeMultiplierUpdate = SlowAdjustingFeeUpdate<Self>;
-    type LengthToFee = ConstantMultiplier<Balance, TransactionByteFee>;
-    type OnChargeTransaction = pallet_transaction_payment::CurrencyAdapter<Balances, ()>;
-    type OperationalFeeMultiplier = OperationalFeeMultiplier;
-    type RuntimeEvent = RuntimeEvent;
-    type WeightToFee = WeightToFee;
-}
-
-impl pallet_sudo::Config for Runtime {
-    type RuntimeCall = RuntimeCall;
-    type RuntimeEvent = RuntimeEvent;
-    type WeightInfo = pallet_sudo::weights::SubstrateWeight<Runtime>;
-}
-
-parameter_types! {
-    pub const ReservedXcmpWeight: Weight = MAXIMUM_BLOCK_WEIGHT.saturating_div(4);
-    pub const ReservedDmpWeight: Weight = MAXIMUM_BLOCK_WEIGHT.saturating_div(4);
-    pub const RelayOrigin: AggregateMessageOrigin = AggregateMessageOrigin::Parent;
-}
-
-type ConsensusHook = cumulus_pallet_aura_ext::FixedVelocityConsensusHook<
-    Runtime,
-    RELAY_CHAIN_SLOT_DURATION_MILLIS,
-    BLOCK_PROCESSING_VELOCITY,
-    UNINCLUDED_SEGMENT_CAPACITY,
->;
-
-impl cumulus_pallet_parachain_system::Config for Runtime {
-    #[cfg(not(feature = "async-backing"))]
-    type CheckAssociatedRelayNumber = RelayNumberStrictlyIncreases;
-    #[cfg(feature = "async-backing")]
-    type CheckAssociatedRelayNumber = RelayNumberMonotonicallyIncreases;
-    type ConsensusHook = ConsensusHook;
-    type DmpQueue = frame_support::traits::EnqueueWithOrigin<MessageQueue, RelayOrigin>;
-    type OnSystemEvent = ();
-    type OutboundXcmpMessageSource = XcmpQueue;
-    type ReservedDmpWeight = ReservedDmpWeight;
-    type ReservedXcmpWeight = ReservedXcmpWeight;
-    type RuntimeEvent = RuntimeEvent;
-    type SelfParaId = parachain_info::Pallet<Runtime>;
-    type WeightInfo = cumulus_pallet_parachain_system::weights::SubstrateWeight<Runtime>;
-    type XcmpMessageHandler = XcmpQueue;
-}
-
-impl parachain_info::Config for Runtime {}
-
-parameter_types! {
-    pub MessageQueueServiceWeight: Weight = Perbill::from_percent(35) * RuntimeBlockWeights::get().max_block;
-    pub const HeapSize: u32 = 64 * 1024;
-    pub const MaxStale: u32 = 8;
-}
-
-impl pallet_message_queue::Config for Runtime {
-    type HeapSize = HeapSize;
-    type IdleMaxServiceWeight = MessageQueueServiceWeight;
-    type MaxStale = MaxStale;
-    #[cfg(feature = "runtime-benchmarks")]
-    type MessageProcessor = pallet_message_queue::mock_helpers::NoopMessageProcessor<
-        cumulus_primitives_core::AggregateMessageOrigin,
-    >;
-    #[cfg(not(feature = "runtime-benchmarks"))]
-    type MessageProcessor = ProcessXcmMessage<
-        AggregateMessageOrigin,
-        xcm_executor::XcmExecutor<xcm_config::XcmConfig>,
-        RuntimeCall,
-    >;
-    // The XCMP queue pallet is only ever able to handle the `Sibling(ParaId)` origin:
-    type QueueChangeHandler = NarrowOriginToSibling<XcmpQueue>;
-    type QueuePausedQuery = NarrowOriginToSibling<XcmpQueue>;
-    type RuntimeEvent = RuntimeEvent;
-    type ServiceWeight = MessageQueueServiceWeight;
-    type Size = u32;
-    type WeightInfo = pallet_message_queue::weights::SubstrateWeight<Runtime>;
-}
-
-impl cumulus_pallet_aura_ext::Config for Runtime {}
-
-parameter_types! {
-    pub const MaxInboundSuspended: u32 = 1000;
-}
-
-impl cumulus_pallet_xcmp_queue::Config for Runtime {
-    type ChannelInfo = ParachainSystem;
-    type ControllerOrigin = EnsureRoot<AccountId>;
-    type ControllerOriginConverter = XcmOriginToTransactDispatchOrigin;
-    type MaxInboundSuspended = MaxInboundSuspended;
-    type PriceForSiblingDelivery = NoPriceForMessageDelivery<ParaId>;
-    type RuntimeEvent = RuntimeEvent;
-    type VersionWrapper = ();
-    type WeightInfo = cumulus_pallet_xcmp_queue::weights::SubstrateWeight<Runtime>;
-    // Enqueue XCMP messages from siblings for later processing.
-    type XcmpQueue = TransformOrigin<MessageQueue, AggregateMessageOrigin, ParaId, ParaIdToSibling>;
-}
-
-parameter_types! {
-    // One storage item; key size is 32; value is size 4+4+16+32 bytes = 56 bytes.
-    pub const DepositBase: Balance = deposit(1, 88);
-    // Additional storage item size of 32 bytes.
-    pub const DepositFactor: Balance = deposit(0, 32);
-    pub const MaxSignatories: u16 = 100;
-}
-
-impl pallet_multisig::Config for Runtime {
-    type Currency = Balances;
-    type DepositBase = DepositBase;
-    type DepositFactor = DepositFactor;
-    type MaxSignatories = MaxSignatories;
-    type RuntimeCall = RuntimeCall;
-    type RuntimeEvent = RuntimeEvent;
-    type WeightInfo = pallet_multisig::weights::SubstrateWeight<Runtime>;
-}
-
-parameter_types! {
-    pub const Period: u32 = 6 * HOURS;
-    pub const Offset: u32 = 0;
-}
-
-impl pallet_session::Config for Runtime {
-    type Keys = SessionKeys;
-    type NextSessionRotation = pallet_session::PeriodicSessions<Period, Offset>;
-    type RuntimeEvent = RuntimeEvent;
-    // Essentially just Aura, but let's be pedantic.
-    type SessionHandler = <SessionKeys as sp_runtime::traits::OpaqueKeys>::KeyTypeIdProviders;
-    type SessionManager = CollatorSelection;
-    type ShouldEndSession = pallet_session::PeriodicSessions<Period, Offset>;
-    type ValidatorId = <Self as frame_system::Config>::AccountId;
-    // we don't have stash and controller, thus we don't need the convert as well.
-    type ValidatorIdOf = pallet_collator_selection::IdentityCollator;
-    type WeightInfo = pallet_session::weights::SubstrateWeight<Runtime>;
-}
-
-#[cfg(not(feature = "async-backing"))]
-parameter_types! {
-    pub const AllowMultipleBlocksPerSlot: bool = false;
-    pub const MaxAuthorities: u32 = 100_000;
-}
-
-#[cfg(feature = "async-backing")]
-parameter_types! {
-    pub const AllowMultipleBlocksPerSlot: bool = true;
-    pub const MaxAuthorities: u32 = 100_000;
-}
-
-impl pallet_aura::Config for Runtime {
-    type AllowMultipleBlocksPerSlot = AllowMultipleBlocksPerSlot;
-    type AuthorityId = AuraId;
-    type DisabledValidators = ();
-    type MaxAuthorities = MaxAuthorities;
-    #[cfg(feature = "async-backing")]
-    type SlotDuration = ConstU64<SLOT_DURATION>;
-    #[cfg(not(feature = "async-backing"))]
-    type SlotDuration = pallet_aura::MinimumPeriodTimesTwo<Self>;
-}
-
-parameter_types! {
-    pub const PotId: PalletId = PalletId(*b"PotStake");
-    pub const SessionLength: BlockNumber = 6 * HOURS;
-    // StakingAdmin pluralistic body.
-    pub const StakingAdminBodyId: BodyId = BodyId::Defense;
-}
-
 /// We allow root and the StakingAdmin to execute privileged collator selection
 /// operations.
 pub type CollatorSelectionUpdateOrigin = EitherOfDiverse<
     EnsureRoot<AccountId>,
     EnsureXcm<IsVoiceOfBody<RelayLocation, StakingAdminBodyId>>,
 >;
-
-parameter_types! {
-    pub const MaxCandidates: u32 = 100;
-    pub const MaxInvulnerables: u32 = 20;
-    pub const MinEligibleCollators: u32 = 4;
-}
-
-impl pallet_collator_selection::Config for Runtime {
-    type Currency = Balances;
-    // should be a multiple of session or things will get inconsistent
-    type KickThreshold = Period;
-    type MaxCandidates = MaxCandidates;
-    type MaxInvulnerables = MaxInvulnerables;
-    type MinEligibleCollators = MinEligibleCollators;
-    type PotId = PotId;
-    type RuntimeEvent = RuntimeEvent;
-    type UpdateOrigin = CollatorSelectionUpdateOrigin;
-    type ValidatorId = <Self as frame_system::Config>::AccountId;
-    type ValidatorIdOf = pallet_collator_selection::IdentityCollator;
-    type ValidatorRegistration = Session;
-    type WeightInfo = pallet_collator_selection::weights::SubstrateWeight<Runtime>;
-}
-
-impl pallet_utility::Config for Runtime {
-    type PalletsOrigin = OriginCaller;
-    type RuntimeCall = RuntimeCall;
-    type RuntimeEvent = RuntimeEvent;
-    type WeightInfo = pallet_utility::weights::SubstrateWeight<Runtime>;
-}
-
-parameter_types! {
-    pub const ProposalBond: Permill = Permill::from_percent(5);
-    pub const ProposalBondMinimum: Balance = 2000; // * CENTS
-    pub const ProposalBondMaximum: Balance = 1;// * GRAND;
-    pub const SpendPeriod: BlockNumber = 6 * DAYS;
-    pub const Burn: Permill = Permill::from_perthousand(2);
-    pub const TreasuryPalletId: PalletId = PalletId(*b"py/trsry");
-    pub const PayoutSpendPeriod: BlockNumber = 30 * DAYS;
-    // The asset's interior location for the paying account. This is the Treasury
-    // pallet instance (which sits at index 13).
-    pub TreasuryInteriorLocation: InteriorLocation = PalletInstance(13).into();
-    pub const MaxApprovals: u32 = 100;
-}
-
-parameter_types! {
-    pub TreasuryAccount: AccountId = Treasury::account_id();
-}
-
-impl pallet_treasury::Config for Runtime {
-    type ApproveOrigin = EitherOfDiverse<EnsureRoot<AccountId>, Treasurer>;
-    type AssetKind = VersionedLocatableAsset;
-    type BalanceConverter = frame_support::traits::tokens::UnityAssetBalanceConversion;
-    #[cfg(feature = "runtime-benchmarks")]
-    type BenchmarkHelper = polkadot_runtime_common::impls::benchmarks::TreasuryArguments;
-    type Beneficiary = VersionedLocation;
-    type BeneficiaryLookup = IdentityLookup<Self::Beneficiary>;
-    type Burn = ();
-    type BurnDestination = ();
-    type Currency = Balances;
-    type MaxApprovals = MaxApprovals;
-    type OnSlash = Treasury;
-    type PalletId = TreasuryPalletId;
-    type Paymaster = PayOverXcm<
-        TreasuryInteriorLocation,
-        crate::xcm_config::XcmRouter,
-        crate::PolkadotXcm,
-        ConstU32<{ 6 * HOURS }>,
-        Self::Beneficiary,
-        Self::AssetKind,
-        LocatableAssetConverter,
-        VersionedLocationConverter,
-    >;
-    type PayoutPeriod = PayoutSpendPeriod;
-    type ProposalBond = ProposalBond;
-    type ProposalBondMaximum = ProposalBondMaximum;
-    type ProposalBondMinimum = ProposalBondMinimum;
-    type RejectOrigin = EitherOfDiverse<EnsureRoot<AccountId>, Treasurer>;
-    type RuntimeEvent = RuntimeEvent;
-    type SpendFunds = ();
-    type SpendOrigin = TreasurySpender;
-    type SpendPeriod = SpendPeriod;
-    type WeightInfo = pallet_treasury::weights::SubstrateWeight<Runtime>;
-}
-
-parameter_types! {
-    pub const PostBlockAndTxnHashes: PostLogContent = PostLogContent::BlockAndTxnHashes;
-}
-
-impl pallet_ethereum::Config for Runtime {
-    type ExtraDataLength = ConstU32<30>;
-    type PostLogContent = PostBlockAndTxnHashes;
-    type RuntimeEvent = RuntimeEvent;
-    type StateRoot = pallet_ethereum::IntermediateStateRoot<Self>;
-}
-
-parameter_types! {
-    pub BlockGasLimit: U256 = U256::from(NORMAL_DISPATCH_RATIO * MAXIMUM_BLOCK_WEIGHT.ref_time() / WEIGHT_PER_GAS);
-    pub GasLimitPovSizeRatio: u64 = BlockGasLimit::get().as_u64().saturating_div(MAX_POV_SIZE);
-    pub WeightPerGas: Weight = Weight::from_parts(WEIGHT_PER_GAS, 0);
-    pub SuicideQuickClearLimit: u32 = 0;
-}
-
-impl pallet_evm::Config for Runtime {
-    type AddressMapping = IdentityAddressMapping;
-    type BlockGasLimit = BlockGasLimit;
-    type BlockHashMapping = pallet_ethereum::EthereumBlockHashMapping<Self>;
-    type CallOrigin = EnsureAccountId20;
-    type ChainId = EVMChainId;
-    type Currency = Balances;
-    type FeeCalculator = BaseFee;
-    type FindAuthor = FindAuthorSession<Aura>;
-    type GasLimitPovSizeRatio = GasLimitPovSizeRatio;
-    type GasWeightMapping = pallet_evm::FixedGasWeightMapping<Self>;
-    type OnChargeTransaction = EVMCurrencyAdapter<Balances, ()>;
-    type OnCreate = ();
-    // FIXME: Will be implemented in #11
-    type PrecompilesType = ();
-    // FIXME: Will be implemented in #11
-    type PrecompilesValue = ();
-    type Runner = pallet_evm::runner::stack::Runner<Self>;
-    type RuntimeEvent = RuntimeEvent;
-    type SuicideQuickClearLimit = SuicideQuickClearLimit;
-    type Timestamp = Timestamp;
-    // FIXME: run the benchmarks
-    type WeightInfo = pallet_evm::weights::SubstrateWeight<Self>;
-    type WeightPerGas = WeightPerGas;
-    type WithdrawOrigin = EnsureAccountId20;
-}
-
-impl pallet_evm_chain_id::Config for Runtime {}
-
-parameter_types! {
-    /// Starting value for base fee. Set at the same value as in Ethereum.
-    pub DefaultBaseFeePerGas: U256 = U256::from(1_000_000_000);
-    /// Default elasticity rate. Set at the same value as in Ethereum.
-    pub DefaultElasticity: Permill = Permill::from_parts(125_000);
-}
-
-/// The thresholds based on which the base fee will change.
-pub struct BaseFeeThreshold;
-impl pallet_base_fee::BaseFeeThreshold for BaseFeeThreshold {
-    fn lower() -> Permill {
-        Permill::zero()
-    }
-
-    fn ideal() -> Permill {
-        Permill::from_parts(500_000)
-    }
-
-    fn upper() -> Permill {
-        Permill::from_parts(1_000_000)
-    }
-}
-impl pallet_base_fee::Config for Runtime {
-    type DefaultBaseFeePerGas = DefaultBaseFeePerGas;
-    type DefaultElasticity = DefaultElasticity;
-    type RuntimeEvent = RuntimeEvent;
-    type Threshold = BaseFeeThreshold;
-}
-
-pub struct FindAuthorSession<F>(PhantomData<F>);
-impl<F: FindAuthor<u32>> FindAuthor<H160> for FindAuthorSession<F> {
-    fn find_author<'a, I>(digests: I) -> Option<H160>
-    where
-        I: 'a + IntoIterator<Item = (ConsensusEngineId, &'a [u8])>,
-    {
-        if let Some(author_index) = F::find_author(digests) {
-            let account_id: AccountId = Session::validators()[author_index as usize];
-            return Some(H160::from(account_id));
-        }
-        None
-    }
-}
-
-#[derive(Clone)]
-pub struct TransactionConverter;
-
-impl fp_rpc::ConvertTransaction<UncheckedExtrinsic> for TransactionConverter {
-    fn convert_transaction(&self, transaction: pallet_ethereum::Transaction) -> UncheckedExtrinsic {
-        UncheckedExtrinsic::new_unsigned(
-            pallet_ethereum::Call::<Runtime>::transact { transaction }.into(),
-        )
-    }
-}
-
-impl fp_rpc::ConvertTransaction<opaque::UncheckedExtrinsic> for TransactionConverter {
-    fn convert_transaction(
-        &self,
-        transaction: pallet_ethereum::Transaction,
-    ) -> opaque::UncheckedExtrinsic {
-        let extrinsic = UncheckedExtrinsic::new_unsigned(
-            pallet_ethereum::Call::<Runtime>::transact { transaction }.into(),
-        );
-        let encoded = extrinsic.encode();
-        opaque::UncheckedExtrinsic::decode(&mut &encoded[..])
-            .expect("Encoded extrinsic is always valid")
-    }
-}
 
 // Create the runtime by composing the FRAME pallets that were previously
 // configured.
