+++ generic-template/node/src/service.rs
@@ -23,6 +23,7 @@ use frame_benchmarking_cli::SUBSTRATE_REFERENCE_HARDWARE;
 use parachain_template_runtime::{
     apis::RuntimeApi,
     opaque::{Block, Hash},
+    configs::TransactionConverter
 };
 use sc_client_api::Backend;
 use sc_consensus::ImportQueue;
